2026-05-30 14:20 CST The issues with cfsecram have been resolved, and with that clean compiling, the rest of 3.1 clean compiled.  The code is finally back into a testable, runnable state.

+++ src/main/java/server/markhome/mcf/v3_1/cfint/cfintram/CFIntRamMimeTypeTable.java
@@ -281,7 +281,7 @@ public class CFIntRamMimeTypeTable
 		ICFIntMimeType iBuff )
 	{
 		CFIntBuffMimeType Buff = (CFIntBuffMimeType)ensureRec(iBuff);
-		Integer pkey = Buff.getPKey();
+		Integer pkey = (Integer)Buff.getPKey();
 		CFIntBuffMimeType existing = dictByPKey.get( pkey );
 		if( existing == null ) {
 			throw new CFLibStaleCacheDetectedException( getClass(),
